feat: delegate smt get max_delegator_size from delegate requirement cell
